Change / rewrite history to use the tools provided by state
This drops the rebase functionality—it was broken and it's not immediately
clear how to repair it, so let's look into that when we are working on an
actual collab implementation.

Fixes the test introduced in efb35d7a447ac6

Stores only mapping data, not full changes, for the maps kept in the history
state.

Makes branches flat arrays.

Plus some other simplifications and optimizations.