Make last TypeScript checks pass
Upgrade w3c-keyname to version with .d.ts file, fix type,
widen lib setting to es2017. The latter isn't great, since we
want to use ES5 in the main sources, but I can't figure out
how to make tsc do that in a single tsconfig.json, so that'll
have to wait until we overhaul the build system.