Revise the way language-attached data is represented
To allow providing it through a facet

BREAKING: Syntax objects no longer have a `languageDataAt` method
(this was moved to `EditorState`, and changed somewhat).

FEATURE: The editor state now has a `languageDataAt` that collects
values stored under a given property name in the main language data
and in object attached for a given document type with the
`addLanguageData` facet.

Closes #140