Add some features to highlighting system
FEATURE: Highlighting information for a syntax (as in `styleTags`) can
now prefix tag names with `!` to make them override the styling of any
child nodes for that syntax node.

FEATURE: It is now possible, when specifying syntax highlighting
information, to write rules for a node that apply only when its parent
nodes are of a given type. This uses `/` syntax, as in
`"Declaration/Identifier": "className"`.