Start on a language-metadata package
Add all the legacy modes to it for a start.

FEATURE: The new language-desc package lists a large number of modes, with
some metadata to help locate the one you're interested in.

FEATURE: The language package now exports a `LanguageDescription` type
that can be used to provide metadata about language packages.